move query msg handling to struct
