+++ contracts/cw-ics721-bridge/src/lib.rs
@@ -5,6 +5,7 @@ pub mod ibc;
 pub mod ibc_helpers;
 pub mod ibc_packet_receive;
 pub mod msg;
+mod query;
 pub mod state;
 pub mod token_types;
